flutter analyze --watch
