+++ dev/bots/test.sh
@@ -45,7 +45,7 @@ SRC_ROOT=$PWD
 
 # generate and analyze our large sample app
 dart dev/tools/mega_gallery.dart
-(cd dev/benchmarks/mega_gallery; flutter watch --benchmark)
+(cd dev/benchmarks/mega_gallery; flutter analyze --watch --benchmark)
 
 if [ -n "$COVERAGE_FLAG" ]; then
   GSUTIL=$HOME/google-cloud-sdk/bin/gsutil
