+++ packages/flutter_tools/lib/executable.dart
@@ -38,7 +38,6 @@ import 'src/commands/test.dart';
 import 'src/commands/trace.dart';
 import 'src/commands/update_packages.dart';
 import 'src/commands/upgrade.dart';
-import 'src/commands/watch.dart';
 import 'src/device.dart';
 import 'src/doctor.dart';
 import 'src/globals.dart';
@@ -85,8 +84,7 @@ Future<Null> main(List<String> args) async {
     ..addCommand(new TestCommand())
     ..addCommand(new TraceCommand())
     ..addCommand(new UpdatePackagesCommand(hidden: !verboseHelp))
-    ..addCommand(new UpgradeCommand())
-    ..addCommand(new WatchCommand(verboseHelp: verboseHelp));
+    ..addCommand(new UpgradeCommand());
 
   return Chain.capture/*<Future<Null>>*/(() async {
     // Initialize globals.
